requirements.txt
#CherryPy==6.0.1
Django==1.8
django-tastypie==0.13.3
numpy==1.11.0
pandas==0.18.1
psutil==4.3.1
python-dateutil==2.5.2
python-mimeparse==1.5.1
pytz==2016.4
requests==2.11.1
django-picklefield==0.3.2
psycopg2==2.6.2
